Updated for B220
NOTE - TournamentMode plugin is default disabled, and doesnt work.  It
doesnt cause errors, but currently does not seem to prevent round start.
May look into this, but probably not an issue as NS2stats has a better
tournament plugin built in now (and best not to cause compat issues if
possible).